requirements.txt
discord.py
numpy
matplotlib
opencv-python-headless
psycopg2